requirements.txt
aiomqtt==2.3.0
pip>=21.3.1
pydantic==2.10.6
PyYAML==6.0.2
ruff==0.9.6
telnetlib3==2.0.4